+++ engine/src/ecs/prefab/sprite.rs
@@ -7,7 +7,7 @@ use crate::utils::{Vec2F, Vec3F};
 
 struct SpriteBuilderState {
     path: String,
-    aspect_ratio: f32,
+    aspect_ratio: f64,
 }
 
 // struct SpriteBuilder;
@@ -21,11 +21,11 @@ struct SpriteBuilderState {
 
 // }
 
-static QUAD_VERTICES: [f32; 32] = [
-    0.5f32, 0.5f32, 0.0f32, 0f32, 0f32, -1f32, 1.0f32, 1.0f32, // top right
-    0.5f32, -0.5f32, 0.0f32, 0f32, 0f32, -1f32, 1.0f32, 0.0f32, // bottom right
-    -0.5f32, -0.5f32, 0.0f32, 0f32, 0f32, -1f32, 0.0f32, 0.0f32, // bottom left
-    -0.5f32, 0.5f32, 0.0f32, 0f32, 0f32, -1f32, 0.0f32, 1.0f32, // top left
+static QUAD_VERTICES: [f64; 32] = [
+    0.5f64, 0.5f64, 0.0f64, 0f64, 0f64, -1f64, 1.0f64, 1.0f64, // top right
+    0.5f64, -0.5f64, 0.0f64, 0f64, 0f64, -1f64, 1.0f64, 0.0f64, // bottom right
+    -0.5f64, -0.5f64, 0.0f64, 0f64, 0f64, -1f64, 0.0f64, 0.0f64, // bottom left
+    -0.5f64, 0.5f64, 0.0f64, 0f64, 0f64, -1f64, 0.0f64, 1.0f64, // top left
 ];
 
 static QUAD_INDICES: [u32; 6] = [0, 1, 2, 2, 3, 0];
